+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- add timeout flag to `mlp deploy` command
+
 ## [v2.2.0] - 2025-09-04
 
 ### Changed